Reland "[llvm][AArch64] Insert "bti j" after call to setjmp"
This reverts commit ea2939bcb8e3016ad04ec823be795372f374bc27.

This changes BLR_BTI to take variable_ops meaning that we can accept
a register or a label. The pattern still expects one argument so we'll
never get more than one. Then later we can check the type of the operand
to choose BL or BLR to emit.

(this is what BLR_RVMARKER does but I missed this detail of it first time around)

Also require NoSLSBLRMitigation which I missed in the first version.