[M68k][test] Remove redundant CHECK-LABEL directive
The associated test had a redundant CHECK-LABEL directive that might fail
the test since the inception, but this issue was "burried" by a missing
colon, which was addressed in 88382eef82c0c245d12789a92d016b7f1181b645.
Thus, the test finally failed after the said commit.

This patch remove that CHECK-LABEL directive.